Restored  ndpi_get_api_version() prototype

+++ src/lib/ndpi_main.c
@@ -6574,7 +6574,7 @@ void NDPI_DUMP_BITMASK(NDPI_PROTOCOL_BITMASK a) {
 }
 #endif
 
-int ndpi_get_api_version() {
+u_int16_t ndpi_get_api_version() {
   return(NDPI_API_VERSION);
 }
